Assignment-06/3.Deploy-liara/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pymssql
sqlalchemy
python-dotenv
pyodbc